moved view related code into its own file and created notify container for all push notifications/popups
